requirements.txt
streamlit
spacy
torch
transformers
gensim
scikit-learn
pdfminer.six
